refactor/1684 parser directory layout
