+++ src/assets/scripts/client/commands/aircraftCommand/aircraftCommandMap.js
