+++ src/assets/scripts/client/commands/parsers/argumentValidators.js
@@ -2,8 +2,8 @@ import _isNaN from 'lodash/isNaN';
 import _isString from 'lodash/isString';
 import _forEach from 'lodash/forEach';
 import { convertStringToNumber } from '../../utilities/unitConverters';
-import { EXPEDITE } from './aircraftCommandMap';
-import { ERROR_MESSAGE } from './aircraftCommandParserMessages';
+import { EXPEDITE } from '../aircraftCommand/aircraftCommandMap';
+import { ERROR_MESSAGE } from './parserMessages';
 import {
     INVALID_INDEX,
     REGEX
