+++ src/assets/scripts/client/commands/scopeCommand/ScopeCommandModel.js
