+++ src/assets/scripts/client/commands/scopeCommand/scopeCommandMap.js
